# src/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DUCC Website API",
    "description": "API documentation for the Durham University Canoe Club website.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie used for authentication."
      }
    }
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user account with the provided email, password, and name details.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "The user's email address." },
                  "password": { "type": "string", "format": "password", "description": "The user's password." },
                  "first_name": { "type": "string", "description": "The user's first name." },
                  "last_name": { "type": "string", "description": "The user's last name." }
                },
                "required": ["email", "password", "first_name", "last_name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Validation failed" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "description": "Authenticates a user and starts a new session.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "User email address." },
                  "password": { "type": "string", "format": "password", "description": "User password." }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Authentication failed" }
        }
      }
    },
    "/api/auth/logout": {
      "get": {
        "summary": "Logout user",
        "description": "Ends the current user session and clears the authentication cookie.",
        "tags": ["Auth"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Logged out" }
        }
      }
    },
...
    "/api/auth/reset-password-request": {
        "post": {
            "summary": "Request password reset",
            "description": "Sends a password reset link to the user's email if the account exists.",
            "tags": ["Auth"],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": { "type": "string", "format": "email", "description": "Email to send reset link to." }
                            },
                            "required": ["email"]
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Reset link sent (if account exists)" },
                "400": { "description": "Email is required" }
            }
        }
    },
    "/api/auth/reset-password": {
        "post": {
            "summary": "Reset password with token",
            "description": "Updates the user's password using a valid reset token from an email link.",
            "tags": ["Auth"],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": { "type": "string", "description": "The password reset token." },
                                "newPassword": { "type": "string", "format": "password", "description": "The new password to set." }
                            },
                            "required": ["token", "newPassword"]
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Password updated successfully" },
                "400": { "description": "Invalid token or missing data" }
            }
        }
    },
    "/api/auth/change-password": {
        "post": {
            "summary": "Change password",
            "description": "Allows an authenticated user to change their current password.",
            "tags": ["Auth"],
            "security": [{ "cookieAuth": [] }],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "currentPassword": { "type": "string", "format": "password", "description": "User's current password." },
                                "newPassword": { "type": "string", "format": "password", "description": "The new password to set." }
                            },
                            "required": ["currentPassword", "newPassword"]
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Password changed successfully" },
                "403": { "description": "Incorrect current password" }
            }
        }
    },
...
  }
}